Defect/Story Number: QC defect #81152 Description Of Defect/Story: OAS-880_LAS Online - 057 – TDC Microphone & Headphone calibration - MAC -  Flash setting popup (only privacy tab) is shown for A/B Client. Who Did Code Review: Soumav Ray Code Files Changes: login.html(OAS98), TestClient.js(OAS98), PersistenceServlet.class(OAS98), PersistenceServlet.java(OAS98), Request.lzx(OAS98) Impacted Areas: "Flash setting pop-up" for A/B and C/D client How The Issue Was Solved:A conditional check has been added so as to show the pop-up only for C/D Windows client and not for A/B(Windows/MAC). Hints To QA For Testing: Schedule a session for for Form A/B test (any level). Login to test client for speaking subtest. Click on Start button. Flash setting pop-up should not appear for Form A/B test client. Developer Comments:	Fix for defect 81152. The behavior of the current QA A/B client is now in sync with the prod A/B client with respect to "Flash setting pop-up" .A conditional check has been added so as to show the pop-up only for C/D Windows client and not for A/B(Windows/MAC).

+++ application-tdc/src/com/ctb/tdc/web/servlet/PersistenceServlet.java
@@ -212,7 +212,7 @@ public class PersistenceServlet extends HttpServlet {
 		else if (method != null && method.equals(ServletUtils.VERIFY_SETTINGS_METHOD)) {
 			result = verifyServletSettings();
 		} else if (method != null && method.equals(ServletUtils.LOGIN_METHOD))
-			result = login(xml);
+			result = login(xml, request);
 		else if (method != null && method.equals(ServletUtils.SAVE_METHOD))
 		{
 			System.out.println(" original save xml:"+xml);			
@@ -458,7 +458,7 @@ public class PersistenceServlet extends HttpServlet {
 	 * 
 	 * @param String xml
 	 */
-	private String login(String xml) {
+	private String login(String xml, HttpServletRequest request) {
 		String result = ServletUtils.ERROR;
 
 		try {
@@ -484,6 +484,11 @@ public class PersistenceServlet extends HttpServlet {
 			}
 
 			File flashCookieTarget = null;
+			//fix for defect 81152. The flash setting popu-up should only appear for C/D MAC client
+			String formType = "false";
+			if(request.getParameter("isCD") != null){
+				formType = request.getParameter("isCD").toString();
+			}
 			if (osName.indexOf("win") >= 0) {
 				
 				boolean successWindows = new File(userHome + windowsPath)
@@ -493,10 +498,13 @@ public class PersistenceServlet extends HttpServlet {
 			} else if (osName.indexOf("mac") >= 0) {
 				new File(userHome + macPath).mkdirs();
 				flashCookieTarget = new File(userHome + macPath	+ "//settings.sol");
-				if (flashCookieTarget.exists())
+				if (flashCookieTarget.exists() && formType != null && formType.equalsIgnoreCase("true"))
 				{
 					boolean successMac = flashCookieTarget.delete();
 					//logger.info(flashCookieTarget.getAbsolutePath()+(successMac?" deleted ":" not deleted "));
+				}else{
+					if (!flashCookieTarget.exists())
+						flashCookieTarget.createNewFile();
 				}
 			} else {
 				boolean successUnix = new File(userHome + unixPath).mkdirs();
@@ -518,15 +526,14 @@ public class PersistenceServlet extends HttpServlet {
 				outFile = new FileOutputStream(flashCookieTarget);
 				byte[] buf = new byte[1024];
 				int len;
-				if (!(osName.indexOf("mac") >= 0)) //don't copy file for mac
-				{
-					//logger.info("Copying file ");
+				if ((!(osName.indexOf("mac") >= 0) && formType != null && formType.equalsIgnoreCase("true"))
+						|| (formType != null && formType.equalsIgnoreCase("false"))){
 					while ((len = in.read(buf)) > 0)
 					{
 						outFile.write(buf, 0, len);
-					}
-
+					}					
 				}
+				
 			}catch (FileNotFoundException e) {
 				logger.info(flashCookieTarget.getAbsolutePath()+" not found");
